requirements.txt
flask==3.1.0
Werkzeug>=3.1.0
mistralai==1.5.1
python-dotenv==1.0.0
python-multipart==0.0.9
requests==2.31.0
markdown==3.6
